+++ modules/Scheduling/AddDrop.php
@@ -9,7 +9,14 @@ $start_date = RequestedDate( 'start', date( 'Y-m' ) . '-01' );
 // Set end date.
 $end_date = RequestedDate( 'end', DBDate() );
 
-echo '<form action="' . PreparePHP_SELF() . '" method="GET">';
+echo '<form action="' . PreparePHP_SELF( [], [
+	'month_start',
+	'day_start',
+	'year_start',
+	'month_end',
+	'day_end',
+	'year_end',
+] ) . '" method="GET">';
 
 DrawHeader(
 	_( 'Timeframe' ) . ': ' . PrepareDate( $start_date, '_start', false ) . ' ' .